testing, remove middleware
